Use CAIID to identify parent image instead of tag
The current "latest" tag corresponds closely to the MARTI production
environment.  If the maintainer upgrades any of those components, we'll
pick up the updates in the next release, which could break our test
environment.  Because there is no other meaningful tag associated with
the current "latest" tag, we have no choice but to reference the parent
image by its CAIID.