Resync with latest production code
Somehow the production code got out of sync with the version used to
initialize this repository.  This commit ensures the repo code is in
sync with the latest production code while maintaining the changes
previously applied to remove hard-coded sensitive information from the
source code, as well as to allow MARTI to run from a Docker container.