Replace trigger_error() with exit()
trigger_error() does not exit the script and continues script execution
on the next line.  This can lead to incorrect behavior.

For example, if the REPLACE SQL statement in register.php fails for some
reason, the script will still send the user a registration email.
Because no record exists in the pending_validations table, all attempts
the user makes to validate their email will fail.  This was the exact
scenario we had when users would attempt to register using an IPv6
address.